tests/wpt/tests/preload/single-download-late-used-preload.html

<!DOCTYPE html>
<title>Ensure preloaded resources are not downloaded again when used</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/preload/resources/preload_helper.js"></script>
<link rel=preload href="resources/square.png?pipe=trickle(d1)" as=image>
<script>
    setup({ single_test: true });
    var link = document.getElementsByTagName("link")[0]
    assert_equals(link.as, "image");
    link.addEventListener("load", () => {
        verifyPreloadAndRTSupport();
        verifyNumberOfResourceTimingEntries("resources/square.png?pipe=trickle(d1)", 1);
        var img = document.createElement("img");
        img.src = "resources/square.png?pipe=trickle(d1)";
        img.onload = () => {
            verifyLoadedAndNoDoubleDownload("resources/square.png?pipe=trickle(d1)");
            done();
        };
        document.body.appendChild(img);
    });
</script>
<body>
